cleanup: silence build warnings
- mark counters/ includes as SYSTEM (suppress external-lib template noise)
- gate now_sec() in pivco_prof.c behind PIVCO_PROF
- mark unused file-format utilities (xxh32, put_u8, get_u8) with
  __attribute__((unused))
- remove dead gain variable in huffman_table.c (math now in comment)
- add (void)simd_n in primitives_neon.h to match the avx512
  PIVCO_PROF=0 path
- file-level pragma in bench_fusion_v3/v4/v5_cnt.cpp for
  unused-lambda-capture noise
- (void)param; / __attribute__((unused)) on isolated bench helpers
- one printf %zu cast in bench_lemire_compare.cpp
- drop the unused c_total_flat aggregate in bench_flat_subtree_stats.c

cmake --build build --clean-first now produces zero warnings.

+++ extras/bench/bench_lemire_compare.cpp
@@ -278,7 +278,7 @@ int main() {
         double ns_call_best = agg.fastest_elapsed_ns() / (double)N_GROUPS;
         double ns_call_avg  = agg.elapsed_ns() / (double)N_GROUPS;
         std::printf("%-30s %10.3f  (avg %.3f, inner_count=%zu)\n",
-                    name, ns_call_best, ns_call_avg, agg.inner_count);
+                    name, ns_call_best, ns_call_avg, (size_t)agg.inner_count);
     };
 
     print_lemire("L1 lemire baseline (vec)",   lemire_baseline(v_src, v_L, v_R, v_m));
